Continue Modelling E.Gadd's Training Room Slide
Replacing BOB for CCM for the E.Gadd training room slide so that it can properly load spindrifts. Modelling the next 1/3rd of E.Gadd's training room including the spindrifts section and the joke room.

Need to model the odyssey dive tutorial area (platforming up around the pillar) and the final part of the slide (which should include an optional path requiring the odyssey dive controls to reach / complete) and the empty painting gallery room.