Adding Inside the Mansion Stage
Adding a stage for the inside of the mansion. Includes a basic slide shape (with fully detailed textures!) and lots of boos to chase the player as they slide.
Boos move about 10x faster now.
Need a couple of things - a big boo in the stage ending which spawns a star in the right location.
A set of red coins and a red coin star.
Some text on the big boo to taunt the player about Luigi being captured at the top of the mansion (can probably just be at the start of the level).
Some coins along the level path.
Custom music!
A more clear opening at the start to get to the slide part of the stage.
Some tiered death planes so you die if you fall off the slide early.